app/requirements.txt
python-docx
psycopg2-binary
uvicorn[standard]
fastapi